+++ providers/common/compat/docs/changelog.rst
@@ -25,6 +25,18 @@
 Changelog
 ---------
 
+1.14.3
+......
+
+Misc
+~~~~
+
+* ``Add OpenLineage parent and transport info injection to 'EmrServerlessStartJobOperator' (#64807)``
+* ``Add OpenLineage parent info injection to GlueJobOperator (#64513)``
+
+.. Below changes are excluded from the changelog. Move them to
+   appropriate section above if needed. Do not delete the lines(!):
+
 1.14.2
 ......